+++ authlete/dto/authorization_fail_reason.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2020 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -31,3 +31,4 @@ class AuthorizationFailReason(Enum):
     ACCOUNT_SELECTION_REQUIRED = auto()
     CONSENT_REQUIRED           = auto()
     INTERACTION_REQUIRED       = auto()
+    INVALID_TARGET             = auto()