+++ authlete/dto/authorization_issue_request.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2020 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,14 +22,15 @@ from authlete.types.jsonable import Jsonable
 class AuthorizationIssueRequest(Jsonable):
     def __init__(self, nameAndValues=None):
         nameAndTypes = {
-            'ticket':      str,
-            'subject':     str,
-            'sub':         str,
-            'authTime':    int,
-            'acr':         str,
-            'claims':      str,
-            'properties':  Property, # list of Property
-            'scopes':      str,      # list of str
+            'ticket':          str,
+            'subject':         str,
+            'sub':             str,
+            'authTime':        int,
+            'acr':             str,
+            'claims':          str,
+            'properties':      Property, # list of Property
+            'scopes':          str,      # list of str
+            'idtHeaderParams': str,
         }
 
         super().__init__(nameAndValues, nameAndTypes)