+++ authlete/dto/authorization_response.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2020 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -45,6 +45,8 @@ class AuthorizationResponse(ApiResponse):
             'requestObjectPayload': str,
             'idTokenClaims':        str,
             'userInfoClaims':       str,
+            'resources':            str,    # list of str
+            'purpose':              str,
             'responseContent':      str,
             'ticket':               str
         }