+++ authlete/dto/backchannel_authentication_fail_reason.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2020 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -25,5 +25,6 @@ class BackchannelAuthenticationFailReason(Enum):
     MISSING_USER_CODE        = auto()
     INVALID_USER_CODE        = auto()
     INVALID_BINDING_MESSAGE  = auto()
+    INVALID_TARGET           = auto()
     ACCESS_DENIED            = auto()
     SERVER_ERROR             = auto()