+++ CHANGES.md
@@ -1,6 +1,10 @@
 CHANGES
 =======
 
+- `AuthorizationFailReason` class
+    * Added `INVALID_TARGET`.
+
+
 1.0.2 (2019-08-09)
 ------------------