+++ CHANGES.md
@@ -17,6 +17,9 @@ CHANGES
 - `BackchannelAuthenticationCompleteResponse` class
     * Added `resources`.
 
+- `BackchannelAuthenticationFailReason` class
+    * Added `INVALID_TARGET`.
+
 
 1.0.2 (2019-08-09)
 ------------------