+++ CHANGES.md
@@ -80,6 +80,10 @@ CHANGES
     * Added `htm`.
     * Added `htu`.
 
+- `TokenResponse` class
+    * Added `accessTokenResources`.
+    * Added `resources`.
+
 
 1.0.2 (2019-08-09)
 ------------------