+++ CHANGES.md
@@ -99,6 +99,7 @@ CHANGES
 - New classes
     * `PushedAuthReqAction`
     * `PushedAuthReqRequest`
+    * `PushedAuthReqResponse`
 
 
 1.0.2 (2019-08-09)