+++ CHANGES.md
@@ -1,6 +1,9 @@
 CHANGES
 =======
 
+1.0.3 (2020-10-01)
+------------------
+
 - `AuthleteApi` class
     * Added `pushedAuthorizationRequest` method.